requirements.txt
absl-py==0.15.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
JIT==0.0.1
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kapre==0.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.3.6
matplotlib==3.3.4
msgpack==1.0.4
msgpack-numpy==0.4.8
numba==0.48.0
numpy==1.19.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.0.3
Pillow==8.4.0
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-ignite==0.3.0
pytz==2021.3
pyzmq==23.2.1
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rir-generator==0.1.0
rsa==4.8
scikit-learn==0.22.1
scipy==1.4.1
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
torchfile==0.1.0
tornado==6.1
tqdm==4.43.0
typing-extensions==3.7.4.3
urllib3==1.26.8
visdom==0.1.8.9
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
